+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enchancements
 
 - Add tri-state boolean editing in the Data Editor with a checkbox control for FALSE, TRUE, and NULL values
+- Add `Drop Schema` context menu command for schema nodes
 
 ## 0.0.3